requirements.txt
pymongo~=4.10.1
watchdog~=4.0.2
flet~=0.25.2
python-dotenv~=1.0.1